transfer some code from server.js to db.js
