+++ devhack2025/db.js
@@ -1,11 +1,27 @@
+const express = require("express");
 const { Pool } = require("pg");
+
+const app = express();
+app.use(express.json());
 
 const pool = new Pool({
   user: "postgres",
-  host: "localhost",
+  host: "localhost", // Change this if your DB is on a remote server
   database: "postgres",
   password: "123456789",
-  port: 5432, // default PostgreSQL port
+  port: 5432,
 });
 
-module.exports = pool;
+// Example Route: Get All Users
+app.get("/users", async (req, res) => {
+  try {
+    const result = await pool.query("SELECT * FROM users");
+    res.json(result.rows);
+  } catch (err) {
+    console.error(err.message);
+    res.status(500).send("Server Error");
+  }
+});
+
+const PORT = 5000;
+app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
