+++ devhack2025/server/server.js
@@ -1,5 +1,3 @@
-//------------------CREATE THE SERVER---------------------
-
 var express = require("express"),
   path = require("path"),
   bodyParser = require("body-parser"),
@@ -29,24 +27,3 @@ app.get("/", function (req, res) {
 app.listen(PORT, function () {
   console.log("Server started on Port 3000");
 });
-
-//--------------SETTING UP EXPRESS API TO INTERACT WITH POSTGRESQL----------------
-
-const express = require("express");
-const pool = require("./db"); // Import database connection
-
-const app = express();
-app.use(express.json()); // Middleware for parsing JSON
-
-// Route to get all users
-app.get("/users", async (req, res) => {
-  try {
-    const result = await pool.query("SELECT * FROM users"); // SQL query
-    res.json(result.rows); // Send result as JSON
-  } catch (err) {
-    console.error(err.message);
-    res.status(500).send("Server Error");
-  }
-});
-
-app.listen(PORT, () => console.log(`Server running on port ${PORT}`));
